+++ .gitignore
@@ -37,4 +37,4 @@ build/
 ### Mac OS ###
 .DS_Store
 
-src/main/resources/application.properties
+database/src/main/resources/application.properties
