+++ .idea/artifacts/web_Web.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="war" name="web:Web">
+    <output-path>$PROJECT_DIR$/out/artifacts/web_Web</output-path>
+    <root id="archive" name="web_Web.war">
+      <element id="artifact" artifact-name="web:Web exploded" />
+    </root>
+  </artifact>
+</component>
