+++ database/src/main/java/com/arinc/entity/Message.java
