--- src/main/web/WEB-INF/regAndLog/styles.css
@@ -1,176 +1,0 @@
-body {
-    margin: 0;
-    padding: 0;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    background-color: #333; /* Dark background color */
-    color: #fff; /* Light text color */
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-}
-
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.form-container {
-    background-color: #444; /* Dark tile background color */
-    color: #fff; /* Light text color */
-    padding: 20px;
-    border-radius: 5px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    animation: fadeIn 0.5s ease-in-out;
-    width: 400px; /* Ширина формы */
-    position: relative;
-}
-form h2 {
-    text-align: center;
-    margin-bottom: 20px;
-    color: #3498db; /* Акцентный цвет заголовка */
-} 
-form label {
-    margin-top: 10px;
-    color: #fff;; /* Цвет текста метки */
-    font-size: 14px;
-}
-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-
-
-
-
-form input {
-    padding: 15px;
-    margin-bottom: 20px;
-    border: none;
-    border-bottom: 2px solid #3498db; /* Линия под инпутом */
-    background-color: #2c3e50; /* Темный цвет инпута */
-    color: #fff; /* Цвет текста инпута */
-    transition: border-color 0.3s ease-in-out;
-}
-
-form input:focus {
-    border-color: #3498db; /* Акцентный цвет границы при фокусе */
-    outline: none;
-}
-
-
-form button {
-    background-color: #666; /* Dark button background color */
-    color: #fff; /* Light text color */
-    margin-left: 10px;
-    height: 50px;
-    width: 100px;
-    padding: 10px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s ease, color 0.3s ease;
-}
-
-form button:hover {
-    background-color: #2980b9; /* Темнее при наведении */
-}
-
-.success-message {
-    text-align: center;
-    color: #27ae60; /* Зеленый цвет для успешных сообщений */
-    font-weight: bold;
-    margin-top: 20px;
-    opacity: 0;
-    max-height: 0;
-    overflow: hidden;
-    transition: max-height 0.3s ease-in-out, opacity 0.3s ease-in-out;
-}
-
-.success-message.show {
-    max-height: 100px;
-    opacity: 1;
-}
-.form-container a, .separator {
-    color: #bdc3c7;
-    font-size: 14px;
-}
-.form-container a:hover {
-    color: #3498db;
-}
-a {
-    color: #fff;;
-    text-decoration: none;
-    transition: color 0.3s ease-in-out;
-}
-
-a:hover {
-    color: #2980b9;
-}
-
-.separator {
-    margin: 20px 0;
-    width: 100%;
-    height: 1px;
-    background-color: #bdc3c7;
-}
-
-.social-buttons {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-}
-
-.social-button {
-    background-color: #3498db;
-    color: #fff;;
-    border: none;
-    border-radius: 50%;
-    padding: 10px;
-    margin: 0 10px;
-    cursor: pointer;
-    transition: background-color 0.3s ease-in-out;
-}
-.forgot-password {
-    text-align: right;
-    margin-top: 10px;
-    color: #bdc3c7;
-    font-size: 14px;
-}
-
-.social-button:hover {
-    background-color: #2980b9;
-}
-
-.forgot-password {
-    text-align: right;
-    margin-top: 10px;
-    color: #bdc3c7;
-}
-
-.error-message {
-    text-align: center;
-    color: #e74c3c; /* Красный цвет для ошибок */
-    font-weight: bold;
-    margin-top: 20px;
-}
-
-.page-title {
-    text-align: center;
-    font-size: 24px;
-    margin-bottom: 20px;
-    color: #ecf0f1;
-}
-
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-    }
-    to {
-        opacity: 1;
-    }
-}
