+++ utils/src/main/java/com/arinc/exceptions/FailedToRegisterDriverException.java
