+++ utils/src/main/java/com/arinc/util/ConnectionManager.java
