+++ utils/src/main/java/com/arinc/util/EntityBuilder.java
