+++ utils/src/main/java/com/arinc/util/PropertiesUtil.java
