+++ web/src/main/webapp/WEB-INF/home/phone-pos.css
@@ -1,0 +1,41 @@
+@media only screen and (max-width: 768px) {
+    .container {
+        grid-template-columns: repeat(3, 1fr);
+        grid-template-rows: repeat(8, 1fr);
+    }
+
+    #chat {
+        grid-column: span 3;
+        grid-row: span 1;
+    }
+    .header-tile.tile{
+        display: block;
+        margin: 0;
+    }
+    #header-buttons{
+        justify-content: center;
+    }
+    .tile{
+        padding: 5px;
+    }
+
+    .online-users-tile.tile{
+        grid-column: span 1;
+        grid-row: span 2;
+    }
+
+    .profile-tile.tile
+    {
+        grid-column: span 2;
+        grid-row: span 1;
+    }
+    .profile-info-tile{
+        grid-column: span 3;
+        grid-row: span 1;
+    }
+    .clock.tile{
+        grid-column: span 2;
+        grid-row: span 1;
+    }
+}
+/*Доделать с телефоном , баг с бесконечным блоком chat*/
