+++ web/src/main/webapp/WEB-INF/home/script.js
