+++ web/src/main/webapp/WEB-INF/home/style-dark.css
@@ -1,0 +1,339 @@
+body {
+    background-color: #333; /* Dark background color */
+    color: #fff; /* Light text color */
+    font-family: 'lucida console', Tahoma, Geneva, Verdana, sans-serif;
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+.container {
+    display: grid;
+    grid-template-columns: repeat(7, 4fr);
+    grid-template-rows: repeat(3, 4fr);
+    gap: 20px;
+    padding: 20px;
+    height: 100%;
+    box-sizing: border-box;
+}
+
+
+.header-tile {
+    height: fit-content;
+    text-align: center;
+   margin:  20px 20px 0 20px;
+   display: flex;
+   flex-direction: row;
+    box-sizing: border-box;
+}
+
+h1 {
+    font-size: 2.5em;
+    margin: 0;
+}
+
+p {
+    font-size: 1.2em;
+}
+
+
+
+.tile {
+    background-color: #444; /* Dark tile background color */
+    color: #fff; /* Light text color */
+    padding: 20px;
+    border-radius: 5px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+}
+
+h2 {
+    font-size: 1.5em;
+}
+
+
+.message {
+    background-color: #555; /* Dark message background color */
+    padding: 10px;
+    margin-bottom: 10px;
+    border-radius: 5px;
+    overflow-wrap: break-word;
+}
+
+#messageInput {
+    flex: 2;
+    padding: 10px;
+    margin-right: 10px;
+    border-radius: 50px;
+    border: none;
+    
+    background-color: #666; /* Dark input background color */
+    color: #fff; /* Light text color */
+}
+
+#messageForm {
+    margin-top: auto;
+    display: flex;
+}
+
+button {
+    background-color: #666; /* Dark button background color */
+    color: #fff; /* Light text color */
+    padding: 12px 20px;
+    margin-left: 10px;
+    border: none;
+    border-radius: 50px; /* Rounded corners */
+    cursor: pointer;
+    transition: background-color 0.3s ease, color 0.3s ease;
+    font-size: 16px;
+    font-weight: bold;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Shadow effect */
+}
+
+.emoji-button {
+    display: flex;
+    
+    
+    border: none;
+    border-radius: 50px; /* Rounded corners */
+    cursor: pointer;
+    transition: background-color 0.3s ease, color 0.3s ease;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Shadow effect */
+}
+
+.emoji-icon {
+    width: 40px; /* Adjust the width as needed */
+    height: 40px;
+    border: none;
+    transition: transform 0.3s ease, border-color 0.3s ease;
+}
+
+
+
+.emoji-button:hover {
+    background-color: #2980b9;
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2); /* Darker shadow on hover */
+}
+
+button:hover {
+    background-color: #2980b9;
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2); /* Darker shadow on hover */
+}
+.clock.tile {
+    text-align: center;
+    grid-column: span 1;
+    grid-row: span 1;
+    
+}
+
+#clockDisplay {
+    font-size: 2em;
+}
+.chat.tile{
+    grid-column: span 5;
+    grid-row: span 2;
+    display: flex;
+    flex-direction: column;
+}
+
+@keyframes slideIn {
+    0% {
+        opacity: 0;
+        transform: translateY(-20px);
+    }
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+.header-tile p {
+    animation: slideIn 1s ease-in-out;
+}
+
+@keyframes slideInRight {
+    0% {
+        opacity: 0;
+        transform: translateX(50px);
+    }
+    100% {
+        opacity: 1;
+        transform: translateX(0);
+    }
+}
+
+.chat .message {
+    animation: slideInRight 0.5s ease-in-out;
+}
+
+@keyframes fadeIn {
+    0% {
+        opacity: 0;
+        transform: translateY(-20px);
+    }
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+h1, h2, h3 {
+    animation: fadeIn 1s ease-in-out;
+}
+
+@keyframes fadeInList {
+    0% {
+        opacity: 0;
+        transform: translateY(-20px);
+    }
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.online-users ul {
+    list-style-type: none;
+    padding: 0;
+    animation: fadeInList 1s ease-in-out;
+}
+
+.profile-tile {
+   
+    grid-row: span 1;
+    grid-column: span 1;
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+}
+
+.profile-tile h2 {
+    font-size: 1.5em;
+    margin-bottom: 10px;
+}
+
+.profile-button {
+    padding: 10px 20px;
+    margin-top: auto;
+    color: #fff;
+    border: none;
+    border-radius: 50px;
+    cursor: pointer;
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.profile-button:hover {
+    background-color: #2980b9;
+}
+
+/* Your existing styles here */
+
+/* Styles for the online users tile */
+.online-users-tile.tile {
+    grid-column: span 1;
+    grid-row: span 2;
+    text-align: center;
+    transition: box-shadow 0.3s ease;
+}
+
+.online-users-tile h2 {
+    font-size: 1.5em;
+    margin-bottom: 15px;
+    
+}
+
+.online-users-list {
+    list-style-type: none;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: flex-start;
+}
+
+.online-users-list li {
+    margin-bottom: 10px;
+    font-size: 1.2em;
+    
+}
+
+.tile:hover {
+    transition: box-shadow 0.3s ease;
+    box-shadow: 0 8px 16px #939393; /* Slightly larger shadow on hover */
+}
+
+.profile-info-tile {
+    grid-column: span 7;
+    padding: 20px;
+    border-radius: 10px; /* Rounded corners */
+    text-align: center;
+}
+
+.profile-info-tile h2 {
+    font-size: 1.5em;
+    margin-bottom: 15px;
+   
+}
+
+.profile-info {
+    text-align: left;
+    margin-bottom: 15px;
+}
+
+.profile-info p {
+    font-size: 1.2em;
+    margin: 8px 0;
+    
+}
+
+.edit-profile-button {
+    padding: 10px 20px;
+    
+    color: #fff;
+    border: none;
+    border-radius: 50px;
+    cursor: pointer;
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.edit-profile-button:hover {
+    background-color: #2980b9; /* Darker shade on hover */
+}
+
+.header-buttons {
+    display: flex;
+    align-items: center;
+    margin-left: auto;
+    justify-content: flex-end;
+}
+
+.theme-switch-button, .login-button {
+    background-color: #666; /* Dark button background color */
+    color: #fff; /* Light text color */
+    margin-left: 10px;
+    height: 50px;
+    width: fit-content;
+    padding: 10px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.theme-switch-button:hover, .login-button:hover {
+    background-color: #2980b9;
+}
+
+.login-button{
+    width: 100px;
+}
+form{
+margin: 0;
+}
+
+.message-box{
+    flex-grow: 1; /* Блок message-box будет занимать все доступное свободное пространство */
+    overflow-y: auto;
+    margin-bottom: 10px;
+    overflow-x: hidden;
+}
+
+
