+++ web/src/main/webapp/WEB-INF/regAndLog/registration.jsp
