+++ web/src/main/webapp/WEB-INF/regAndLog/styles.css
@@ -1,0 +1,137 @@
+/* Общие стили для всего документа */
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+/* Шрифт для текста */
+body {
+    font-family: Arial, sans-serif;
+    font-size: 16px;
+    color: white;
+    background-color: #222;
+}
+
+/* Контейнер для центрирования содержимого */
+.container {
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 20px;
+    background: linear-gradient(to right, #222, #333, #444);
+}
+
+/* Стили для формы */
+.form-container {
+    border: 1px solid #444;
+    border-radius: 10px;
+    padding: 40px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+
+/* Заголовок формы */
+h2 {
+    text-align: center;
+    margin-bottom: 20px;
+}
+
+/* Стили для полей ввода */
+input {
+    display: block;
+    width: 100%;
+    height: 40px;
+    border: none;
+    outline: none;
+    padding: 10px 10px 10px 40px;
+    margin-bottom: 10px;
+    background-color: #333;
+    color: white;
+    transition: 0.3s;
+    background-repeat: no-repeat;
+    background-position: 10px center;
+}
+input:focus {
+    transform: scale(1.1);
+    box-shadow: inset 0 0 5px white;
+}
+
+/* Стили для кнопки отправки */
+button {
+    display: block;
+    width: 100%;
+    height: 50px;
+    border: none;
+    outline: none;
+    cursor: pointer;
+    margin-top: 20px;
+    background-color: #555;
+    color: white;
+    transition: 0.3s;
+    transform: perspective(500px) rotateX(-10deg);
+    padding-left: 40px;
+    background-repeat: no-repeat;
+    background-position: 10px center;
+}
+
+/* Эффект наведения на кнопку */
+button:hover {
+    background-color: #666;
+
+    transform: perspective(500px) rotateX(0deg);
+}
+
+/* Стили для сообщения об ошибке */
+div[style] {
+    border: 1px solid red;
+    border-radius: 5px;
+}
+
+/* Стили для ссылки на регистрацию */
+p {
+    text-align: center;
+    margin-top: 20px;
+}
+
+/* Стили для заголовка страницы */
+.page-title {
+    font-size: 24px;
+}
+
+/* Стили для сообщения об успешной регистрации */
+.success-message {
+    display: none;
+    text-align: center;
+    font-size: 20px;
+    font-weight: bold;
+    color: green;
+}
+
+/* Стили для разделителя */
+.separator {
+    height: 1px;
+    background-color: #444;
+    margin: 20px 0;
+}
+
+/* Стили для ссылки на вход */
+.login-link {
+    text-align: center;
+}
+
+/* Стили для анимации кнопки */
+.button-pulse {
+    animation: pulse 2s infinite;
+}
+
+/* Ключевые кадры для анимации */
+@keyframes pulse {
+    0% {
+        transform: scale(1);
+    }
+    50% {
+        transform: scale(1.1);
+    }
+    100% {
+        transform: scale(1);
+    }
+}
